+++ src/dactyl_keyboard/param/tree/nested.clj
@@ -9,7 +9,7 @@
 
 (def raws
   "A flat version of a special part of a user configuration."
-  [["# Nestable configuration options\n\n"
+  [["# Nestable key configuration options\n\n"
     "This document describes all those settings which can be made at any "
     "level of specificity, from the entire keyboard down to one side of one "
     "key. These settings all go under `by-key` in a YAML file, as indicated "
@@ -19,21 +19,24 @@
     "\n"
     "The `by-key` section contains a map of up to five items:\n"
     "\n"
-    "- `parameters`, where you put actual settings for the entire keyboard. "
+    "- `parameters`, where you put your settings. "
     "Sections described in this document all pertain to this map.\n"
     "- `clusters`, starting a nested map for specific clusters only, "
     "keyed by their names.\n"
     "- `columns` and/or `rows`, each starting a nested map for specific "
     "columns or rows only, keyed either by their indices (ordinal integers) "
-    "or by the special words `first` or `last`. Due to a peculiarity of the "
-    "YAML parser, **numeric indices must appear in quotation marks** as in the "
-    "example below.\n"
-    "- `sides`, starting a nested map for specific sides only, "
-    "keyed by the long-form cardinal points of the compass, i.e. the words "
-    "`north`, `east`, `south` or `west`.\n"
+    "or by the special words `first` or `last`. Due to a peculiarity of "
+    "YAML (and JSON), **numeric indices must appear in quotation marks** "
+    "as in the example below.\n"
+    "- `sides`, starting a nested map for points of the compass, such as "
+    "`NNE` for the north by northeast corner. Sides are only relevant for "
+    "the periphery of a key, including details of the wall surrounding a "
+    "key mount at the edge of a cluster, and the webbing between keys.\n"
     "\n"
     "Each of the nested maps have the same structure as this root-level map. "
-    "Greater specificity is accomplished by nesting a series of these maps.\n"
+    "Specificity is accomplished by nesting a series of these maps, so that "
+    "a nested set of `parameters` comes to refer to an intersection of more "
+    "than one selection criterion.\n"
     "\n"
     "### Example\n"
     "\n"
@@ -41,7 +44,8 @@
     "Once at the root level and twice with enough selection criteria to limit "
     "the effect to two individual keys.\n"
     "\n"
-    "```by-key:\n"
+    "```\n"
+    "by-key:\n"
     "  parameters:\n"
     "    key-style: plump\n"
     "  clusters:\n"
@@ -59,12 +63,27 @@
     "In this example, `key-style` will have the value `plump` for all keys "
     "except two. It will have the value `svelte` for the key closest to the "
     "user (first row) and the key three steps above the home row (row 3), both "
-    "in the second column from the left (column 1) of key cluster `C`.\n"
+    "in the second column from the left (i.e. column 1) of key cluster `C`.\n"
     "\n"
     "Key cluster `C` and the two key styles must be defined elsewhere. Also, "
     "if the keyboard uses reflection, notice that the descriptions given in "
-    "the previous paragraph must be mirrored for the left-hand side of the "
-    "keyboard.\n"
+    "the previous paragraph would be incorrect for the left-hand side of the "
+    "keyboard, because that side would be mirrored.\n"
+    "\n"
+    "### Comparison to anchoring\n"
+    "\n"
+    "The `by-key` section can place any key anywhere. However, please don’t. "
+    "Look first to the `key-clusters` section, which is intended to define "
+    "the most basic shape and position of each cluster. "
+    "`key-clusters` allows for standard [anchoring](options-anchoring.md), "
+    "while `by-key` does not.\n"
+    "\n"
+    "Please use `key-clusters` to translate and rotate each cluster. "
+    "More concretely, prefer `key-clusters` over `translation` and over "
+    "base `pitch`, `roll` and `yaw` at a global or cluster level of `by-key`. "
+    "Doing so is easier, because standard anchoring allows you to describe "
+    "rotation on all three axes together, while keeping values in `by-key` "
+    "smaller and neatly unentangled.\n"
     "\n"
     "### Unique selection criteria\n"
     "\n"
@@ -149,112 +168,149 @@
     "- No cluster, no column, row -1, no side.\n"
     "- No cluster, no column, no row, west side.\n"
     "- No cluster, no column, no row, no side.\n"]
-   [:section [:layout]
+   [[:key-style]
+    {:default :default :parse-fn keyword}
+    "The name of a key style defined in the [global](options-main.md) `keys` "
+    "section. The default setting is the name `default`."]
+   [[:layout]
     "Settings for how to place keys."]
-   [:section [:layout :matrix]
+   [[:layout :matrix]
     "Roughly how keys are spaced out to form a matrix."]
-   [:section [:layout :matrix :neutral]
+   [[:layout :matrix :neutral]
     "The neutral point in a column or row is where any progressive curvature "
     "both starts and has no effect."]
-   [:parameter [:layout :matrix :neutral :column]
+   [[:layout :matrix :neutral :column]
     {:default 0 :parse-fn int}
     "An integer column ID."]
-   [:parameter [:layout :matrix :neutral :row]
+   [[:layout :matrix :neutral :row]
     {:default 0 :parse-fn int}
     "An integer row ID."]
-   [:section [:layout :matrix :separation]
+   [[:layout :matrix :separation]
     "Tweaks to control the systematic separation of keys. The parameters in "
     "this section will be multiplied by the difference between each affected "
     "key’s coordinates and the neutral column and row."]
-   [:parameter [:layout :matrix :separation :column]
+   [[:layout :matrix :separation :column]
     {:default 0 :parse-fn num}
     "A distance in mm."]
-   [:parameter [:layout :matrix :separation :row]
+   [[:layout :matrix :separation :row]
     {:default 0 :parse-fn num}
     "A distance in mm."]
-   [:section [:layout :pitch]
+   [[:layout :pitch]
     "Tait-Bryan pitch, meaning the rotation of keys around the x axis."]
-   [:parameter [:layout :pitch :base]
+   [[:layout :pitch :base]
     {:default 0 :parse-fn parse/compass-incompatible-angle}
-    "An angle in radians. Set at cluster level, this controls the general "
-    "front-to-back incline of the key cluster."]
-   [:parameter [:layout :pitch :intrinsic]
+    "An angle in radians, controlling a uniform front-to-back incline."]
+   [[:layout :pitch :intrinsic]
     {:default 0 :parse-fn parse/compass-incompatible-angle}
     "An angle in radians. Intrinsic pitching occurs early in key placement. "
     "It is typically intended to produce a tactile break between two rows of "
     "keys, as in the typewriter-like terracing common on flat keyboards with "
     "OEM-profile or similarly angled caps."]
-   [:parameter [:layout :pitch :progressive]
+   [[:layout :pitch :progressive]
     {:default 0 :parse-fn parse/compass-incompatible-angle}
     "An angle in radians. This progressive pitch factor bends columns "
     "lengthwise. If set to zero, columns are flat."]
-   [:section [:layout :roll]
+   [[:layout :roll]
     "Tait-Bryan roll, meaning the rotation of keys around the y axis."]
-   [:parameter [:layout :roll :base]
+   [[:layout :roll :base]
     {:default 0 :parse-fn parse/compass-incompatible-angle}
-    "An angle in radians. This is the “tenting” angle. Applied to your main "
-    "cluster on a split keyboard, it controls the overall left-to-right tilt "
-    "of each half."]
-   [:parameter [:layout :roll :intrinsic]
+    "An angle in radians, controlling a uniform right-to-left incline, also "
+    "known as tenting."]
+   [[:layout :roll :intrinsic]
     {:default 0 :parse-fn parse/compass-incompatible-angle}
     "An angle in radians, analogous to intrinsic pitching. Where more than "
     "one column of keys is devoted to a single finger at the edge of the "
     "keyboard, this can help make the edge column easier to reach, reducing "
     "the need to bend the finger (or thumb) sideways."]
-   [:parameter [:layout :roll :progressive]
+   [[:layout :roll :progressive]
     {:default 0 :parse-fn parse/compass-incompatible-angle}
     "An angle in radians. This progressive roll factor bends rows "
     "lengthwise, which also gives the columns a lateral curvature."]
-   [:section [:layout :yaw]
+   [[:layout :yaw]
     "Tait-Bryan yaw, meaning the rotation of keys around the z axis."]
-   [:parameter [:layout :yaw :base]
+   [[:layout :yaw :base]
     {:default 0 :parse-fn parse/compass-incompatible-angle}
-    "An angle in radians. Applied to your main key cluster, this serves the "
-    "purpose of allowing the user to keep their wrists straight even if the "
-    "two halves of the keyboard are closer together than the user’s shoulders."]
-   [:parameter [:layout :yaw :intrinsic]
+    "An angle in radians, corresponding to the way your hand naturally "
+    "describes an arc as you rotate your arm horizontally at the elbow. "
+    "Yawing columns of keys can allow the user to keep their wrists straight "
+    "even on a keyboard shorter than the width of the user’s own shoulders."]
+   [[:layout :yaw :intrinsic]
     {:default 0 :parse-fn parse/compass-incompatible-angle}
     "An angle in radians, analogous to intrinsic pitching."]
-   [:section [:layout :translation]
+   [[:layout :translation]
     "Translation in the geometric sense, displacing keys in relation to each "
     "other. Depending on when this translation takes places, it may have a "
     "a cascading effect on other aspects of key placement. All measurements "
     "are three-dimensional vectors in mm."]
-   [:parameter [:layout :translation :early]
+   [[:layout :translation :early]
     {:default [0 0 0] :parse-fn vec :validate [::tarmi-core/point-3d]}
     "”Early” translation happens before other operations in key placement and "
     "therefore has the biggest knock-on effects."]
-   [:parameter [:layout :translation :mid]
+   [[:layout :translation :mid]
     {:default [0 0 0] :parse-fn vec :validate [::tarmi-core/point-3d]}
     "This happens after columns are styled but before base pitch and roll. "
     "As such it is a good place to adjust whole columns for relative finger "
     "length."]
-   [:parameter [:layout :translation :late]
+   [[:layout :translation :late]
     {:default [0 0 0] :parse-fn vec :validate [::tarmi-core/point-3d]}
     "“Late” translation is the last step in key placement and therefore "
     "interacts very little with other steps."]
-   [:parameter [:key-style]
-    {:default :default :parse-fn keyword}
-    "The name of a key style defined in the [global](options-main.md) `keys` "
-    "section. The default setting is the name `default`."]
-   [:section [:channel]
+   [[:layout :clearance]
+    "The height of each key above its mounting plate can be included when "
+    "calculating the effective radius of key cluster curvature for a layout. "
+    "This height is defined as the distance between the lower edge of a keycap "
+    "and the top of the mounting plate, when the switch is at rest. Clearance "
+    "is defined this way because caps are presumed to be widest at the lower "
+    "edge and therefore most likely to collide at this height."]
+   [[:layout :clearance :use-key-style]
+    {:default false :parse-fn boolean}
+    "If `true`, predict clearance based on `key-style`, including switch "
+    "travel and configured keycap skirt length. "
+    "Ironically, this is most practical when the keyboard design "
+    "is closely tied to the key style. Depending on how other features of the "
+    "design are anchored, clearance based on key style can introduce knock-on "
+    "effects that make it difficult to adapt the design to other switches."]
+   [[:layout :clearance :nominal]
+    {:default 0 :parse-fn num}
+    "Nominal clearance in mm. This is only used without `use-key-style`."]
+   [[:channel]
     "Above each switch mount, there is a channel of negative space for the "
     "user’s finger and the keycap to move inside. This is only useful in those "
     "cases where nearby walls or webbing between mounts on the keyboard would "
     "otherwise obstruct movement."]
-   [:parameter [:channel :height]
+   [[:channel :height]
     {:default 1 :parse-fn num}
     "The height in mm of the negative space, starting from the "
     "bottom edge of each keycap in its pressed (active) state."]
-   [:parameter [:channel :top-width]
+   [[:channel :top-width]
     {:default 0 :parse-fn num}
     "The width in mm of the negative space at its top. Its width at the "
     "bottom is defined by keycap geometry."]
-   [:parameter [:channel :margin]
+   [[:channel :margin]
     {:default 0 :parse-fn num}
     "The width in mm of extra negative space around the edges of a keycap, on "
     "all sides. This is applied before the `error-general` DFM compensator."]
-   [:section [:wall]
+   [[:plate]
+    "The properties of the flat mounting plate through which each switch is "
+    "inserted."]
+   [[:plate :use-key-style]
+    {:default false :parse-fn boolean}
+    "If `true`, base the size and position of the plate on `key-style`."]
+   [[:plate :size]
+    {:default [1 1 1] :parse-fn vec :validate [::tarmi-core/point-3d]}
+    "Three measurements in mm. If `use-key-style` is `false`, this is the "
+    "size of the mounting plate."]
+   [[:plate :position]
+    {:default [0 0 0] :parse-fn vec :validate [::tarmi-core/point-3d]}
+    "A three-dimensional offset in mm. If `use-key-style` is `false`, the "
+    "mounting plate is automatically lowered underneath "
+    "the centre of the switch. This offset is applied at the same time.\n"
+    "\n"
+    "This parameter is intended for adapting the mounting plate as such to "
+    "exotic human interfaces. To adjust the position of the entire key for "
+    "other purposes, use the `translation` settings for `layout`, in this "
+    "document."]
+   [[:wall]
     "Properties of a wall built around the edge of the cluster.\n"
     "\n"
     "The walls of the keyboard case support the key mounts and protect the "
@@ -263,13 +319,8 @@
     "\n"
     "The `wall` section determines the shape of the case wall, specifically "
     "the skirt around each key mount along the edges of the board. These "
-    "skirts are made up of convex hulls wrapping sets of corner posts.\n"
-    "\n"
-    "There is one corner post at each actual corner of every key mount "
-    "(segment 0). "
-    "More posts are displaced from it, going down the sides. Their placement "
-    "is affected by the way the key mounts are rotated etc.\n"]
-   [:parameter [:wall :thickness]
+    "skirts are made up of convex hulls wrapping sets of corner posts."]
+   [[:wall :thickness]
     {:default [1 1 1] :parse-fn parse/pad-to-3-tuple
      :validate [::tarmi-core/point-3d]}
     "The size in mm of the key mount and each wall post.\n\n"
@@ -286,29 +337,114 @@
     "`tweaks`. Whereas mounting-plate thickness ignores the x- and y-axis "
     "dimensions of this parameter, wall posts are cuboids that use all three "
     "dimensions."]
-   [:parameter [:wall :extent]
+   [[:wall :extent]
     {:default 0 :parse-fn num :validate [(set (range 4))]}
     "A segment ID describing how far away from the key mount to extend its "
-    "wall. Note that even if this is set low, you can still use `tweaks` to "
-    "target other segments."]
-   [:parameter [:wall :to-ground]
+    "wall. Note that even if this is set lower than the number of segments "
+    "you’ve defined, you can still use `tweaks` to target other segments."]
+   [[:wall :to-ground]
     {:default false :parse-fn boolean}
     "If `true`, draw one extra, vertical section of wall between the segment "
-    "identified in `extent` and the ground beneath the key."]
-   [:parameter [:wall :bevel]
-    {:default 0 :parse-fn num}
-    "A distance in mm, describing where to place some vertical segments.\n"
+    "identified by `extent` and the ground beneath it."]
+   [[:wall :segments]
+    {:default {0 [0 0 0]}
+     :parse-fn (parse/map-of parse/integer (parse/tuple-of num))
+     :validate [:three/segments]}
+    "A map of segment IDs to xyz-coordinates in mm.\n"
     "\n"
-    "The `bevel` is applied at the top of a wall, making up the difference "
-    "between wall segments 0 and 1. It is applied again at the bottom, making "
-    "up the difference between segments 2 and 3. It affects all "
-    "coordinates. The mathematical operation by which it is applied to the z "
-    "coordinate is determined by the sign of `perpendicular`."]
-   [:parameter [:wall :parallel]
-    {:default 0 :parse-fn num}
-    "A distance in mm. Wall segments 2 and 3 extend this far "
-    "away from the corners of their key mount, on its plane."]
-   [:parameter [:wall :perpendicular]
-    {:default 0 :parse-fn num}
-    "A distance in mm. Wall segments 2 and 3 extend this far "
-    "away from the corners of their key mount, along its normal."]])
+    "This map is indexed by wall segment IDs, which are non-negative "
+    "integers. As with column IDs under `columns`, they must be entered in "
+    "YAML as strings.\n"
+    "\n"
+    "The values of the map are three-dimensional offsets. "
+    "Any offset given for segment 0 is relative to a switch mounting plate. "
+    "The default value for segment 0 is `[0, 0, 0]`, which means that walls "
+    "will start to build out from the corner of each mounting plate.\n"
+    "\n"
+    "Segments other than 0, starting with segment 1, are offset relative to "
+    "the preceding segment and have no default value built into the "
+    "application.\n"
+    "\n"
+    "Offsets are *cumulative*. Segments form a chain, each one positioned "
+    "relative to the one before, as the building blocks of each wall.\n"
+    "\n"
+    "Consider this example:\n"
+    "\n"
+    "```\n"
+    "by-key:\n"
+    "  parameters:\n"
+    "    wall:\n"
+    "      extent: 2\n"
+    "      segments:\n"
+    "        "1": [0, 1, -0.5]\n"
+    "        "2": [0, 0, -4]\n"
+    "  sides\n"
+    "    SSE:\n"
+    "      parameters:\n"
+    "        wall:\n"
+    "          segments:\n"
+    "            "2": [0, 0, -10]\n```"
+    "\n\n"
+    "With this configuration, walls will be built connecting segments 0, 1 "
+    "and 2 on the edge of each key cluster. For the sake of illustration, "
+    "Let’s say there’s only one cluster of keys: A, B, and C, in one row. "
+    "Imagine their corners radiating numbered wall segments.\n"
+    "\n"
+    "```\n"
+    "  2–––2–2–––2–2–––2\n"
+    " /1–––1–1–––1–1–––1\\\n"
+    "210–––0–0–––0–0–––012\n"
+    "||| A     B     C |||\n"
+    "210–––0–0–––0–0–––012\n"
+    " \\1–––1–1–––1–1–––1/\n"
+    "  2–––2–2–––2–2–––2\n```"
+    "\n\n"
+    "A more detailed ASCII diagram of the B key names the sides from which "
+    "its wall segments radiate:\n"
+    "\n"
+    "```\n"
+    "–2–––––2–\n"
+    "–1–––––1–\n"
+    "–0–––––0–\n"
+    "NNW   NNE\n"
+    "\n"
+    "   B\n"
+    "\n"
+    "SSW   SSE\n"
+    "–0–––––0–\n"
+    "–1–––––1–\n"
+    "–2–––––2–\n```"
+    "\n\n"
+    "To understand the effect of the example configuration, it is simplest to "
+    "begin thinking about the `NNE` side of the B key in this image. `NNE` "
+    "here does not mean “facing 22½° east of nominal north from the middle of "
+    "the mounting plate”. It means “facing nominal north from the `NE` corner”. "
+    "Segment 0 on the `NNE` side of the B key is positioned precisely at the "
+    "`NE` corner because the example configuration leaves segment 0 at its "
+    "default offset: `[0, 0, 0]`.\n"
+    "\n"
+    "In the vector space of key B, segment 1 is located at `[0, 1, -0.5]` away "
+    "from segment 0. This is 1 mm to the north and ½ mm closer to the ground: "
+    "A minor bevel.\n"
+    "\n"
+    "In the vector space of key B, segment 2 is located at `[0, 1, -4.5]` "
+    "away from segment 0. "
+    "Its `-4` on the z-axis is added to the offset for segment 1, placing it "
+    "directly below segment 1 rather than radiating on the xy-plane alone.\n"
+    "\n"
+    "For other sides of the B key, segment coordinates stated in the "
+    "configuration are flipped and rotated. For example, segment 1 of the "
+    "`SSE` corner is `[0, -1, -4.5]` away from segment 0 on the same corner. "
+    "Segment 2 of the `SSW` corner has a side-specific override noted in the "
+    "example configuration, dipping an extra 6 mm to `[0, -1, -10.5]` away "
+    "from segment 0 on the `SSW` corner.\n"
+    "\n"
+    "The key idea here is that offsets are expressed in the vector space "
+    "local to the key’s nominal *northeast quadrant*. They are used as given "
+    "only for the `NNE` and `NE` sides of the key. Even then, they are later "
+    "subjected to all the same transformations as the mounting plate itself.\n"
+    "\n"
+    "Offsets are *automatically adapted* to each side to make it easy to "
+    "specify something like a bevel with a single value in the configuration, "
+    "without having to state explicitly that the bevel should sweep around "
+    "the key cluster, “turning” as the wall turns."]])